refactor /node/ structure
