+++ src/_node/file/actions.ts
@@ -1,0 +1,5 @@
+export const f_create = () => {};
+export const f_remove = () => {};
+export const f_duplicate = () => {};
+export const f_move = () => {};
+export const f_copy = () => {};
