+++ src/_node/file/index.ts
@@ -2,3 +2,4 @@ export * from "./types";
 export * from "./apis";
 export * from "./helpers";
 export * from "./handlers";
+export * from "./actions";
