+++ src/_node/index.ts
@@ -1,4 +1,5 @@
 export * from "./types";
+export * from "./helpers";
 export * from "./apis";
 export * from "./file";
-export * from "./html";
+export * from "./node";
