+++ src/_node/node/actions.ts
@@ -1,0 +1,5 @@
+export const n_create = () => {};
+export const n_remove = () => {};
+export const n_duplicate = () => {};
+export const n_move = () => {};
+export const n_copy = () => {};
