+++ src/_node/node/type/index.ts
@@ -1,0 +1,1 @@
+export * from "./html";
