+++ src/components/main/actionsPanel/workspaceTreeView/hooks/useNodeViewState.ts
@@ -2,7 +2,7 @@ import { useCallback, useContext } from "react";
 
 import { useDispatch, useSelector } from "react-redux";
 
-import { getValidNodeUids } from "@_node/apis";
+import { getValidNodeUids } from "@_node/helpers";
 import { TNodeUid } from "@_node/types";
 import { MainContext } from "@_redux/main";
 import {
