+++ src/components/main/codeView/hooks/useEditor.tsx
@@ -1,35 +1,60 @@
-import { useCallback, useContext, useEffect, useRef, useState } from "react";
+import {
+  useCallback,
+  useContext,
+  useEffect,
+  useRef,
+  useState,
+} from 'react';
 
-import { debounce } from "lodash";
-import { editor, IPosition } from "monaco-editor";
-import morphdom from "morphdom";
-import * as parse5 from "parse5";
-import { useDispatch, useSelector } from "react-redux";
+import { debounce } from 'lodash';
+import {
+  editor,
+  IPosition,
+} from 'monaco-editor';
+import morphdom from 'morphdom';
+import * as parse5 from 'parse5';
+import {
+  useDispatch,
+  useSelector,
+} from 'react-redux';
 
-import { DefaultTabSize, RootNodeUid } from "@_constants/main";
-import { getSubNodeUidsByBfs } from "@_node/apis";
-import { parseFile, TFileNodeData } from "@_node/file";
+import {
+  DefaultTabSize,
+  RootNodeUid,
+} from '@_constants/main';
+import {
+  parseFile,
+  TFileNodeData,
+} from '@_node/file';
 import {
   PreserveRnbwNode,
   StageNodeIdAttr,
-} from "@_node/file/handlers/constants";
-import { TNode, TNodeTreeData, TNodeUid } from "@_node/types";
-import { MainContext } from "@_redux/main";
-import { setCodeEditing, setCodeViewTabSize } from "@_redux/main/codeView";
+} from '@_node/file/handlers/constants';
+import { getSubNodeUidsByBfs } from '@_node/helpers';
+import {
+  TNode,
+  TNodeTreeData,
+  TNodeUid,
+} from '@_node/types';
+import { MainContext } from '@_redux/main';
+import {
+  setCodeEditing,
+  setCodeViewTabSize,
+} from '@_redux/main/codeView';
 import {
   currentFileUidSelector,
   fileTreeSelector,
   setDoingFileAction,
   setFileTree,
-} from "@_redux/main/fileTree";
+} from '@_redux/main/fileTree';
 import {
   focusNodeTreeNode,
   selectNodeTreeNodes,
   setNodeTree,
-} from "@_redux/main/nodeTree";
-import { setCurrentFileContent } from "@_redux/main/nodeTree/event";
+} from '@_redux/main/nodeTree';
+import { setCurrentFileContent } from '@_redux/main/nodeTree/event';
 
-import { CodeSelection } from "../types";
+import { CodeSelection } from '../types';
 
 function getLanguageFromExtension(extension: string) {
   if (extension) return extension;
