+++ src/components/main/processor/hooks/useProcessorNodeTree.ts
@@ -1,13 +1,16 @@
-import { useContext, useEffect } from "react";
+import {
+  useContext,
+  useEffect,
+} from 'react';
 
-import { useDispatch } from "react-redux";
+import { useDispatch } from 'react-redux';
 
-import { RootNodeUid } from "@_constants/main";
-import { getSubNodeUidsByBfs } from "@_node/apis";
-import { TNodeTreeData } from "@_node/types";
-import { MainContext } from "@_redux/main";
-import { setValidNodeTree } from "@_redux/main/nodeTree";
-import { useAppState } from "@_redux/useAppState";
+import { RootNodeUid } from '@_constants/main';
+import { getSubNodeUidsByBfs } from '@_node/helpers';
+import { TNodeTreeData } from '@_node/types';
+import { MainContext } from '@_redux/main';
+import { setValidNodeTree } from '@_redux/main/nodeTree';
+import { useAppState } from '@_redux/useAppState';
 
 export const useProcessorNodeTree = () => {
   const dispatch = useDispatch();
